% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\name{amidst_data_stream}
\alias{amidst_data_stream}
\title{Creates an AMIDST data stream from an arff file}
\usage{
amidst_data_stream(file_name)
}
\arguments{
\item{file_name}{the name of the data file in arff format}
}
\value{
a Java object of class \code{DataStream}
}
\description{
Creates an AMIDST data stream from an arff file
}

