% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\name{amidst_ds_iterator}
\alias{amidst_ds_iterator}
\title{Creates an iterator over an AMIDST data stream}
\usage{
amidst_ds_iterator(data_on_disk_reader)
}
\arguments{
\item{data_on_disk_reader}{an AMIDST data stream}
}
\value{
an iterator over the stream
}
\description{
Creates an iterator over an AMIDST data stream
}

