% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\name{generate_stream_from_dbn}
\alias{generate_stream_from_dbn}
\title{Generates a stream from a DBN.
The DBN is assumed to be a classifier,
and therefore there is a class variable.}
\usage{
generate_stream_from_dbn(net, sequences, seq_length, class_var)
}
\arguments{
\item{net}{the network that will be sampled}

\item{sequences}{the number of random sequences to generate. Each one will
be an element of the stream.}

\item{seq_length}{the length of the sequences}

\item{class_var}{the class variable}
}
\value{
a Java object of class DataStream with the generated sequences
}
\description{
Generates a stream from a DBN.
The DBN is assumed to be a classifier,
and therefore there is a class variable.
}

