% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\name{generate_stream_of_sequences}
\alias{generate_stream_of_sequences}
\title{Generates a stream where each item is a full sequence from some
dynamic evidence.}
\usage{
generate_stream_of_sequences(dynamic_evidence)
}
\arguments{
\item{dynamic_evidence}{some dynamic observation over s DBN}
}
\value{
a data stream out of the input evidence
}
\description{
Generates a stream where each item is a full sequence from some
dynamic evidence.
}

