% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_amidst_bn}
\alias{load_amidst_bn}
\title{Loads an AMIDST BN from a fle}
\usage{
load_amidst_bn(file_name)
}
\arguments{
\item{file_name}{the name of the file containing the network in .net format}
}
\value{
a java object of class \code{BayesianNetwork}
}
\description{
Loads an AMIDST BN from a fle
}
\examples{
network <- load_amidst_bn(system.file("extdata","WasteIncinerator.bn",
package="ramidst"))
print_amidst_bn(network)
}

