% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_dynamic_amidst_bn}
\alias{load_dynamic_amidst_bn}
\title{Loads an AMIDST dynamic BN from a fle}
\usage{
load_dynamic_amidst_bn(file_name)
}
\arguments{
\item{file_name}{the name of the file containing the network in .net format}
}
\value{
a java object of class \code{DynamicBayesianNetwork}
}
\description{
Loads an AMIDST dynamic BN from a fle
}

