% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_dynamic_sequence}
\alias{load_dynamic_sequence}
\title{Loads an AMIDST dynamic data sequence from a fle}
\usage{
load_dynamic_sequence(file_name)
}
\arguments{
\item{file_name}{the name of the file containing the data sequence}
}
\value{
a Java object of class \code{DataStream}
}
\description{
Loads an AMIDST dynamic data sequence from a fle
}

