% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_inference.R
\name{map_inference_from_stream}
\alias{map_inference_from_stream}
\title{Runs MAP inference from an AMIDST data stream}
\usage{
map_inference_from_stream(network, map_variables, evidence_variables,
  input_stream, sample_size, parallel = T, seed = 3L)
}
\arguments{
\item{network}{a java object of class \code{BayesianNetwork} over which the
computations will be carried out}

\item{map_variables}{a vector with the name of the variables over which the
MAP configuration will be computed}

\item{evidence_variables}{a vector with the names of the observed variables}

\item{input_stream}{and AMIDST data stream}

\item{sample_size}{the sample size to be used for estimating marginals}

\item{parallel}{a \code{boolean} indicating whether or not the items in the
sample will be generated in parallel (when allowed by the system)}

\item{seed}{the seed for the genertion of random numbers}
}
\value{
a \code{data.frame} with the MAP configuration of the
variables of interest for each item in the stream
}
\description{
Runs MAP inference from an AMIDST data stream
}
\note{
The function computes the MAP configuration of the variables of
interest given some evidence for all the items in the input stream.
}
\examples{
\dontrun{
network <- load_amidst_bn(system.file("extdata","WasteIncinerator.bn",
package="ramidst"))
sample_stream <- amidst_data_stream(system.file("extdata",
"WasteIncineratorSample.arff",package="ramidst"))
map_configurations <- map_inference_from_stream(network,c("D","B"),c("W"),
sample_stream,5L)
map_configurations
}
}
\references{
D. Ramos-Lopez, A. Salmeron, R. Rumi, A.M.
Martinez, T.D. Nielsen, A.R. Masegosa, H. Langseth, A.L.
Madsen (2016) Scalable MAP inference in Bayesian networks based on a
Map-Reduce approach. PGM'2016. JMLR: Workshop and Conference Proceedings,
vol. 52: 415-425.
}

