% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{print_amidst_bn}
\alias{print_amidst_bn}
\title{Prints an AMIDST model}
\usage{
print_amidst_bn(network)
}
\arguments{
\item{network}{the network to print out.}
}
\description{
Prints an AMIDST model
}
\examples{
network <- load_amidst_bn(system.file("extdata","WasteIncinerator.bn",
package="ramidst"))
print_amidst_bn(network)
}

