% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramidst.R
\docType{package}
\name{ramidst}
\alias{ramidst}
\alias{ramidst-package}
\title{ramidst: A package for accessing inference and concept drift detection
features provided by the AMIDST Toolbox for handling massive data streams.
Visit \url{http://www.amidsttoolbox.com} for additional information on
the AMIDST Toolbox. This package was partially developed as part of the
AMIDST and PGMs-SDA projects. AMIDST has received funding from the European
Union's 7th Framework Programme for research, technological development and
demonstration under grant no. 619209. PGMs-SDA has received funding from the
Spanish Ministry of Economy and Competitiveness and FEDER funds under grant
TIN2013-46638-C3-1-P.}
\description{
ramidst: A package for accessing inference and concept drift detection
features provided by the AMIDST Toolbox for handling massive data streams.
Visit \url{http://www.amidsttoolbox.com} for additional information on
the AMIDST Toolbox. This package was partially developed as part of the
AMIDST and PGMs-SDA projects. AMIDST has received funding from the European
Union's 7th Framework Programme for research, technological development and
demonstration under grant no. 619209. PGMs-SDA has received funding from the
Spanish Ministry of Economy and Competitiveness and FEDER funds under grant
TIN2013-46638-C3-1-P.
}

