% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{save_dynamic_amidst_bn}
\alias{save_dynamic_amidst_bn}
\title{Saves an AMIDST dynamic BN to a fle}
\usage{
save_dynamic_amidst_bn(network, file_name)
}
\arguments{
\item{network}{a java object of class \code{DynamicBayesianNetwork}}

\item{file_name}{the name of the file that will contain the network}
}
\description{
Saves an AMIDST dynamic BN to a fle
}
\examples{
network <- dbn_generator(1,2,2)
save_dynamic_amidst_bn(network,"sample_net.bn")
new_network <- load_dynamic_amidst_bn("sample_net.bn")
print_amidst_bn(network)
}

